pipeline/requirements.txt
bs4
requests
pylint
pytest
pytest-mock
python-dotenv
lxml
psycopg2-binary
opencage
pandas
